Chapter016/requirements.txt
asgiref==3.8.1
attrs==25.3.0
autopep8==2.3.2
Django==5.1.1
django-extensions==3.2.3
django-filter==25.1
django-silk==5.3.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
gprof2dot==2025.4.14
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
pillow==10.4.0
pycodestyle==2.13.0
PyJWT==2.9.0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.1
sqlparse==0.5.1
tomli==2.2.1
typing_extensions==4.13.2
tzdata==2024.2
uritemplate==4.1.1
